% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depLT-functions.R
\name{positivity.binary}
\alias{positivity.binary}
\title{Diagnostics of non-overlap in the TVW estimator for a binary covariate}
\usage{
positivity.binary(Z, X, T, delta, ylim = NA, xlim = NA)
}
\arguments{
\item{Z}{a binary covariate for which we want to test whether the TVW-specific positivity is not violated.}

\item{X}{a vector of observed (possibly right-censored) lifetimes.}

\item{T}{a vector of truncation times.}

\item{delta}{a censoring indicator. It is assumed that 0 corresponds to censoring.}

\item{censoring}{censoring can get one of the three values "after" (default), "before", "none".}
}
\value{
The function plots the time-varying proportions of 1 and does not return a value.
}
\description{
Visual test for the marginal (i.e. for a single covariate) violation of
positivity (overlap).
The diagnostics are done by visual comparing between the proportions of 1
in a binary covariate \code{Z} corresponding to
observations at risk at each event time, for TVW versus CW estimators.
This test can identify practical, TVW-specific violation of positivity.
But it cannot test theoretical violation of positivity.
}
\examples{
# under construction

}
