% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depLT-functions.R
\name{positivity.cont}
\alias{positivity.cont}
\title{Diagnostics of non-overlap in the TVW estimator for a continuous covariate}
\usage{
positivity.cont(Z, X, T, delta, chosen = NA, ylim = NA, xlim = NA)
}
\arguments{
\item{Z}{a continuous covariate for which we want to test whether the TVW-specific positivity is not violated.}

\item{X}{a vector of observed (possibly right-censored) lifetimes.}

\item{T}{a vector of truncation times.}

\item{delta}{a censoring indicator. It is assumed that 0 corresponds to censoring.}

\item{censoring}{censoring can get one of the three values "after" (default), "before", "none".}
}
\value{
The function plots the distributions and doe not return a value.
}
\description{
Visual test for the marginal (i.e. for a single covariate) violation of
positivity (overlap).
The diagnostics are done by visual comparing between the distributions
of a continuous covariate corresponding to
observations at risk at each event time, for TVW versus CW estimators.
This test can identify practical, TVW-specific violation of positivity.
But it can neither test theoretical violation of positivity, nor
the non-positivity with respect to the lifetime \code{X}  itself.
}
\examples{
# under construction

}
