% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depLT-functions.R
\name{survfit.TVW}
\alias{survfit.TVW}
\title{Time-varying-weight estimation of survival probability}
\usage{
survfit.TVW(X, T, delta, Z, censoring = "after", bs = FALSE, nbs.rep = 200)
}
\arguments{
\item{X}{a vector of observed (possibly right-censored) lifetimes.}

\item{T}{a vector of truncation times.}

\item{delta}{a censoring indicator. It is assumed that 0 corresponds to censoring.}

\item{Z}{a matrix of covariates (common causes of \code{T} and \code{X}) that will be used for estimation
of selection probabilities. NOTE: so far implemented only for one covariate.}

\item{censoring}{censoring can get one of the three values "after" (default), "before", "none".}

\item{bs}{a logical value indicating whether to perform bootstap for estimation of
a confidence interval, and a SE. The default if FALSE.
If bs=FALSE there will not be provided any estimate of uncertainty.}

\item{nbs.rep}{number of bootstrap replications. The default is 200.}

\item{conf.int}{The confidence level for confidence intervals and hypotheses tests.
The default level is 0.95.}

\item{seed}{for the bootstrap}
}
\value{
A list with components:
\tabular{llr}{
\code{time} \tab a vector of time points at which the survival curve has a step \tab   \cr
\code{surv} \tab the estimate of survival at time \code{time}+0. \tab  \cr
\code{SE} \tab the estimate of SE \tab   \cr
\code{CI.L} \tab the pointwise  of a lower limit of the confidence interval \tab  \cr
\code{CI.U} \tab the pointwise  of an upper limit of the confidence interval\tab  \cr}
}
\description{
The estimator assumes the Cox proportional hazards model for truncation time given covariates, which is
used to obtain "time-varying weights" and reweigh the contribution of the observations at-risk accordingly.
For further details we refer to the paper of Vakulenko-Lagun et al. (2020).
}
\examples{
# under construction

}
\references{
Vakulenko-Lagun, B. Qian, J. Chiou, S.-H. Wang, N. Betensky, R.A. 2020. Estimation under covariate-induced dependent truncation using inverse probability weighting. Submitted.
}
